+ requests feature; + email services; change publish novels into services
