+++ routes/routes.py
@@ -11,9 +11,14 @@ router = Router()
 # Novels EN
 router.post("/novels/latest/en", novels_en.publish_latest)
 
+# Routes for requests
 router \
     .post("/novels/requests", novels_requests.requests) \
     .get("/novels/requests", novels_requests.get_all_requests)
+
+# Routes for publish requests
+router \
+    .post("/novels/requests/publish", novels_requests.publish_requests)
 
 # Routes for languages
 router \
